Add files via upload
T-test using Pandas and Scipy to check if having minifigures affect the price of Lego.

H0: The mean of Lego price that has minifigs is equal to that of does not have minifigs.  

H1: The mean of Lego price that has minifigs is not equal to that of does not have minifigs.  

As the result shows, I applied independent equal variance t-test on the price of Lego group by if they have minifigs(=0) or not(>0). The result showed that p-value is larger than 0.05, and we failed to reject the null hypothesis. In short, whether the Lego has minifigs or not does not have impacts on the price of Lego.